// src/types/Shikimori/Responses/Types/AnimeType.ts
import {StatusType} from "@/types/Shikimori/Responses/Types/StatusType";
import {IncompleteDateType} from "@/types/Shikimori/Responses/Types/IncompleteDateType";
import {PosterType} from "@/types/Shikimori/Responses/Types/PosterType";
import {GenreType} from "@/types/Shikimori/Responses/Types/GenreType";
import {StudioType} from "@/types/Shikimori/Responses/Types/StudioType";
import {ExternalLinkType} from "@/types/Shikimori/Responses/Types/ExternalLinkType";
import {PersonRoleType} from "@/types/Shikimori/Responses/Types/PersonRoleType";
import {CharacterRoleType} from "@/types/Shikimori/Responses/Types/CharacterRoleType";
import {RelatedType} from "@/types/Shikimori/Responses/Types/RelatedType";
import {VideoType} from "@/types/Shikimori/Responses/Types/VideoType";
import {ScreenshotType} from "@/types/Shikimori/Responses/Types/ScreenshotType";
import {ScoreStatType} from "@/types/Shikimori/Responses/Types/ScoreStatType";
import {StatusStatType} from "@/types/Shikimori/Responses/Types/StatusStatType";
import {AnimeKindEnum} from "@/types/Shikimori/Responses/Enums/AnimeKindEnum";
import {AnimeRatingEnum} from "@/types/Shikimori/Responses/Enums/AnimeRatingEnum";
import {TopicType} from "@/types/Shikimori/Responses/Types/TopicType";
import {UserRateType} from "@/types/Shikimori/Responses/Types/UserRateType";

export type AnimeType = {
    airedOn: IncompleteDateType | null;
    characterRoles: CharacterRoleType[];
    createdAt: string;
    description: string | null;
    descriptionHtml: string | null;
    descriptionSource: string | null;
    duration: number | null;
    english: string | null;
    episodes: number;
    episodesAired: number;
    externalLinks: ExternalLinkType[];
    fandubbers: string[];
    fansubbers: string[];
    franchise: string | null;
    genres: GenreType[];
    id: string;
    isCensored: boolean | null;
    japanese: string | null;
    kind: AnimeKindEnum | null;
    licenseNameRu: string | null;
    licensors: string[];
    malId: string | null;
    name: string;
    nextEpisodeAt: string | null;
    personRoles: PersonRoleType[];
    poster: PosterType | null;
    rating: AnimeRatingEnum | null;
    related: RelatedType[];
    releasedOn: IncompleteDateType | null;
    russian: string | null;
    score: number | null;
    scoresStats: ScoreStatType[];
    screenshots: ScreenshotType[];
    season: string | null;
    status: StatusType | null;
    statusesStats: StatusStatType[];
    studios: StudioType[];
    synonyms: string[];
    topic: TopicType | null;
    updatedAt: string;
    url: string;
    userRate: UserRateType | null;
    videos: VideoType[];
}